fix(manager): format RBAC subject from integer user id (#4877)
The RBAC middleware built the Casbin enforcement subject with
fmt.Sprint on the id read from the JWT claims. Because JWT claims are
JSON, the id decodes as a float64, and fmt.Sprint renders any float64
value >= 1000000 in scientific notation (for example 1.234567e+06).

The grouping policies are written from the integer user id (a uint), so
the scientific-notation subject never matched any policy and Enforce
returned false. Every request from a user whose id is >= 1000000 was
rejected with HTTP 401 on every manager API endpoint.

Convert the id back to uint before formatting, matching how the audit
middleware already reads the same claim, so the subject equals the
policy subject across the full id range.

Add a regression test that drives the RBAC handler with ids around the
scientific-notation boundary.

+++ manager/middlewares/rbac.go
@@ -49,7 +49,7 @@ func RBAC(e *casbin.Enforcer) gin.HandlerFunc {
 			return
 		}
 
-		if ok, err := e.Enforce(fmt.Sprint(id.(float64)), permission, action); err != nil {
+		if ok, err := e.Enforce(fmt.Sprint(uint(id.(float64))), permission, action); err != nil {
 			logger.Errorf("RBAC validate error: %s", err)
 			c.JSON(http.StatusUnauthorized, gin.H{
 				"message": "permission validate error!",
